[INJI-543] moved styles into style theme files
