+++ components/ui/themes/DefaultTheme.ts
@@ -1,7 +1,14 @@
 /* eslint-disable sonarjs/no-duplicate-string */
-import {Dimensions, I18nManager, StyleSheet, ViewStyle} from 'react-native';
+import {
+  Dimensions,
+  I18nManager,
+  StatusBar,
+  StyleSheet,
+  ViewStyle,
+} from 'react-native';
 import {Spacing} from '../styleUtils';
 import {isIOS} from '../../../shared/constants';
+import Constants from 'expo-constants';
 
 const Colors = {
   Black: '#000000',
@@ -196,7 +203,6 @@ export const DefaultTheme = {
       backgroundColor: Colors.Grey6,
       borderRadius: 4,
     },
-    cardDetailsContainer: {},
     bottomTabIconStyle: {
       padding: 4,
       width: Dimensions.get('window').width * 0.12,
@@ -618,6 +624,18 @@ export const DefaultTheme = {
       borderBottomColor: Colors.Grey5,
       borderBottomWidth: 1,
       marginTop: 10,
+    },
+    introSliderHeader: {
+      marginTop: isIOS()
+        ? Constants.statusBarHeight + 40
+        : StatusBar.currentHeight + 40,
+      width: '100%',
+      marginBottom: 50,
+    },
+    introSliderButton: {
+      borderRadius: 10,
+      height: 50,
+      marginTop: -10,
     },
   }),
   QrCodeStyles: StyleSheet.create({
